kieu double bi loi khong ro lido

+++ Bai8TinhKeThua/PhanB/Bai2HeThongPTGiaoThong/PhuongTien.java
@@ -2,19 +2,22 @@ package Bai8TinhKeThua.PhanB.Bai2HeThongPTGiaoThong;
 public class PhuongTien {
     protected String hangSanXuat;
     protected int namSanXuat;
-    protected double giaBan;
+    protected int giaBan;
 
-    public PhuongTien(String hang, int nam, double gia) {
+    public PhuongTien(String hang, int nam, int gia) {
         this.hangSanXuat = hang;
         this.namSanXuat = nam;
         this.giaBan = gia;
     }
     public String getHangSanXuat() {
-         return hangSanXuat; }
+         return hangSanXuat; 
+        }
     public int getNamSanXuat() { 
-        return namSanXuat; }
-    public double getGiaBan() { 
-        return giaBan; }
+        return namSanXuat;
+     }
+    public int getGiaBan() { 
+        return giaBan; 
+    }
     public int layVanTocToiDa() {
       return 0;
     }
